Add XML Schmas From TS102034
Downloaded all ZIP files for TS102034 available from the website, and
created symlinks (removing version designations from file names) for the
latest versions.

BUT, there are six DTDs contained in the 1.3.1 and 1.4.1 zip files, which
need to have the public and/or system ID set manually in the catalog.xml
(see bug 2301).