Straw-Man Proposal for the Static Page's Structure
This is a straw-man proposal for a document structure of the static web
page. It is pre-populated with (mildly adapted) text from TR 102 679.
Links won't work yet, of course.

H2 headings will get an entry in some navigation bar at the top of the
page. A permalink to the symlink for the latest vesrion of each file is
provided, and a table showing all versions, and their relation to ETSI
spec versions.