Replace Namespace by URI
The first column was cunningly labeled "namespace", because that is what
it is for XML Schema. But in reality, and more generally, it is an
indentifier for the file. In the case of XML Schema it's a urn, and in the
case of JSON Schema it's a http url that identifies it. So to be more
generic the first column should be labeled "URI". That way, we could
easily use the same table scheme for JSON Schema files.