Implement 2017-08-01 Telechat Agreements
Filled in text for the "fixme" comments at the top.

Updated the MHP section with the dual structure as agreed in the telechat
earlier today: one subsection by spec version, and one subsection by file.
Each subsection has an introductory paragraph, which contains a link to
the respective sibling section.

In the by-spec-version subsections, for each version, links to the spec
document and the accompanying ZIP file, both on the ETSI website, are given.
A table lists all files associated with the respective version. The link
on the URI points to the file itself, whereas the "file info" link points
to the part of the by-file subsection that deals with the file.

In the by-file subesction, the lin kon the URI also points top the file
itself. The link on the ETSI spec version points to the section for that
version in the by-spec-version subsection above.

The MHP section is a complete example. In the other setcions, only the
top-level headings and boilerplate pragraphs have been filled in.