Complete IPTV Section
The BCGType classification scheme shipped with the IPTV apecs is
identical with the generic one, so I didn't list it under IPTV, but
instead point to the subsection in the generic part from the "file info"
links for it in the IPTV section.

Also made sure all @href to metadata files on the server are absolute,
i.e. start with a slash as in "/metadata/...". This is to make sure that
wherever the static page will end up on the website, or be moved to later,
the links will never break.