Bug 2874
The system ID of the MHP application description DTD was marked up as
a link, which was probably due to some editor auto-magic. The link is
removed now.

Also, said DTD file was missing from the XML catalogues. To be able to
add it to the catalogue via catbuilder.pl, it needs to sit in a
sub-directory. As it appeared with the first version of the spec, I
moved it into the 1.1.1 sub-directory, and created a symbolic link for
the latest version.